Sort of compiling split verifier
